fixed `v-repeat` directive where collection has 1
If there is only 1 item in the collection, don't bother checking for other elements. `nextEl` does not exists and causes an exception.

I assume this would not affect a change in the model where a new collection also has a size of one, but it's a different object.  As the `vm.$reused` would not have been set?