allow for object definitions as filter arguments
I would like to propose a change to the filter token regex. This will allow for object definitions to be passed as a single argument to the filter. This will only work for shallow objects. For deep objects we would have to use a better parser.

e.g.:

{{ '%count% Pull Requests' | trans {count: count} }}

currently only this works:

{{ '%count% Pull Requests' | trans {count:count} }}