Add support for v-bind:content shorthand (fixes #2843)
This commit aims to add support for `v-bind:content` shorthand. With
this, v-bind with an empty expression will imply the expression to be
the directive name, i.e. `v-bind :content` will have the same effect as
`v-bind:content="content"`, `:href` will have the same effect as
`:href="href"` and so on so forth.

Only attribute bindings are affected by this change. Event, transition,
class/style bindings remain unchanged.