Adjust tests
When checking for undefined components, use 'undefined' instead of a
tag with an empty string, since an empty string will now failover to a
<div> tag. This may sound awkward, but an empty string tag previously
could arguably be considered an error/invalid. By allowing this to
failover to a <div> tag, it makes constructs such as '.active@menu' to
naturally be resolved as <div class="active" id="menu">, etc.