Allow for custom rule of blocks parsing
I am currently working on a Nativescript+Vue template, which includes a specific webpack config that allows for differentiated templates based on the target (web or native). Single file components are extremely useful and being able to add a custom rule to the parsing of their blocks would make it possible to write a single file containing a `native` template alongside a `web` template, both of which sharing the same <script> logic or <style> rules. So far, the only workaround is to change the 'vue-loader' in webpack but by adding this it would be possible to use the actual official 'vue-loader' and change only the necessary part.