fix(slots): preserve reactivity inside of $slots
Previously $slots was not reactive which caused use cases such as
portal-vue to not update when the contents of the slots updated. This
was because there were no reactive properties accessed when using $slots
and renders were only being triggered by calls to $forceUpdate when slot
contents changed.

This change makes $slots reactive, so updates are detected appropriately
regardless of where the slot is used.

fixes #7856