refactor("matches" function inside keep-alive component): Removed duplicated call "indexOf(name) > -
Function accepts three types of argument: string, regexp or array.
If we check that arg is regexp and return after this, we can
make second check for array. If this is not array - we don't need to
call third check, because of we are using flow. It's possible just to
return simple expression.