build(build error message): Add the path of the resolved vue package to the "package version mismatc
In the case that, when building the vue-template-compiler package, the resolved vue package is found
to have a different version than the vue-template-compiler package, add the full path of the
resolved vue package to the resulting error message. This will allow developers to quickly determine
where the problematic vue package is located on their machine.