perf(lifecycle): Prevent rerender of child components that use $listeners
Child components that use the $listeners attribute are rerendered each time the parent component is
updated. This is due to the updateChildComponent function that sets the reactive property
vm.$listeners thus triggering a rerender of the child component. The problem is that we cannot check
if the listeners were updated as we cannot compare functions. This change therefore provides a new
toggleNotifyingSubscribers method that disables a reactive property's ability to notify its
subscribers. We use this method to prevent the vm.$listeners property from rerendering the child
component.

fix #7257