fix(TypeScript): Force ArrayConstructor to be a any[] typescript type
When using an Array prop in a Vue component with TypeScript, typescript breaks because the
JavaScript arrayconstructor is transpiled to a {}[] typescript type that is for some reason more
general than any[]. Since the type is more general, the component is not using the correct component
overload function declaration. The idea here is to force the JavaScript ArrayConstructor to be
transpiled to any[] type.

fix #9935