$mount can be called on comment node
The typescript definition currently specifies that you can only mount on an Element (or a selector string evaluating to an element), however the result of a server-rendered component can be an empty comment node (e.g. if the component has a v-if on the root node). $mount actually supports this, but the type definition does not reflect this currently.