fix(ssr): prevent rendering duplicate css references
There is a chance that the same css file is included in initial and async array, causing a <link>
element pointing to the same stylesheet to be included multiple times. This fix ensures that cannot
happen. Since multiple build systems exist, this defensive approach avoids any issues with how files
are generated regardless of which one is being used by a given project.