feat(core): add prop name to the property validator "invalid prop type" warning
Previously, the "Invalid prop type" warning message returned by the property validator did not
include the name of the offending property. In certain scenarios this obscurity can make it
difficult to identify the source of the issue. To address this, I have added the name of the
offending property to the output as suggested in #12454.

fix #12454