Allow inferring types from Vue.set(object, ...) pattern
Right now it's not possible to infer the type when doing Vue.set this is because Vue assumes it to be a generic object. But with a small change we can allow these types to be inferred in a clear and consistent way similar to how it's array inference works.